feat: implement triple activation
Implement "activation" of triples as part of their life cycle. A triple
can now be either active or inactive, and store now has versions of
it's Get API that only retrieves active triples. The intent is for the
verifier to use these new API.

This decouples adding endorsements to the store from making them
available for verification, allowing of more control of when each
happens, and to maintain an "archive" of no longer active entries.

The add command has been amended with a flag that allows activating
triples as the CoRIMs are added.

Signed-off-by: setrofim <setrofim@gmail.com>