feat(model): handle cached extensions
Update dependencies to the corim library that now caches unknown
extensions. Extract cached extension values along side known extensions
when converting.

An ExtensionValue represents a cached value if its FieldName is empty.
In that case, JSONTag is set to the map key (which could be the
stringified CBOR code point) ValueBytes are set to the CBOR-encoded
value.

Signed-off-by: setrofim <setrofim@gmail.com>